further adaptions to new HW version

+++ software/eclipse/electronicLoad/src/loadFunctions.h
@@ -43,11 +43,11 @@ struct {
 
     struct {
         uint32_t current;
-        uint32_t currentSum;
+        uint64_t currentSum;
         uint32_t voltage;
-        uint32_t voltageSum;
+        uint64_t voltageSum;
         uint32_t power;
-        uint32_t powerSum;
+        uint64_t powerSum;
         uint16_t temp1;
         uint16_t temp2;
         uint32_t nsamples;
